+++ packages/extension-polkagate/src/hooks/useProfiles.ts
@@ -6,6 +6,7 @@ import { AccountContext } from '../components';
 import type { AccountJson } from '@polkadot/extension-base/background/types';
 
 type Profiles = {
+  accountProfiles: string[];
   userDefinedProfiles: string[];
   defaultProfiles: string[];
 };
@@ -14,48 +15,49 @@ export default function useProfiles(account?: AccountJson): Profiles {
   const { accounts } = useContext(AccountContext);
 
   return useMemo(() => {
-     if (!accounts) {
+    if (!accounts) {
       return {
-        userDefinedProfiles: [],
-        defaultProfiles: []
+        accountProfiles: [],
+        defaultProfiles: [],
+        userDefinedProfiles: []
       };
     }
-   
+
     // default profiles
     const defaultProfiles = ['All'];
     const hasLocal = accounts.find(({ isExternal }) => !isExternal)
     if (hasLocal) {
-     defaultProfiles.push('Local')
+      defaultProfiles.push('Local')
     }
 
     const hasLedger = accounts.find(({ isHardware }) => isHardware)
     if (hasLedger) {
-     defaultProfiles.push('Ledger')
+      defaultProfiles.push('Ledger')
     }
 
     const hasWatchOnly = accounts.find(({ isExternal, isQR, isHardware }) => isExternal && !isQR && !isHardware);
     if (hasWatchOnly) {
-     defaultProfiles.push('Watch-only')
+      defaultProfiles.push('Watch-only')
     }
 
     const hasQrAttached = accounts.find(({ isQR }) => isQR);
     if (hasQrAttached) {
-     defaultProfiles.push('QR-attached')
+      defaultProfiles.push('QR-attached')
     }
 
-    if (account) { // return only the account profiles as userDefinedProfiles
-      const profiles = account?.profile?.split(',');
+    let accountProfiles: string[] = [];
+    if (account) {
+      const profiles = account.profile?.split(',');
 
-      return {
-        defaultProfiles,
-        userDefinedProfiles: profiles?.sort() || []
-      };
+      accountProfiles = profiles ? profiles.sort() : [];
+
     }
 
     // user defined profiles
     const profiles = accounts.map(({ profile }) => profile ? profile.split(',') : undefined).flat().filter(Boolean) as string[];
 
     return {
+      accountProfiles,
       defaultProfiles,
       userDefinedProfiles: [...new Set(profiles)].sort()
     };
